Refactor and add asset node menu as context menu
