+++ js_modules/dagster-ui/packages/ui-core/src/asset-graph/AssetNode.tsx
@@ -3,6 +3,7 @@ import {Box, Colors, FontFamily, Icon, Spinner, Tooltip} from '@dagster-io/ui-co
 import countBy from 'lodash/countBy';
 import isEqual from 'lodash/isEqual';
 import React from 'react';
+import ReactDOM from 'react-dom';
 import {Link} from 'react-router-dom';
 import styled, {CSSObject} from 'styled-components';
 
@@ -13,11 +14,13 @@ import {StaleReasonsTags} from '../assets/Stale';
 import {assetDetailsPathForKey} from '../assets/assetDetailsPathForKey';
 import {AssetComputeKindTag} from '../graph/OpTags';
 import {AssetCheckExecutionResolvedStatus, AssetCheckSeverity} from '../graphql/types';
+import {ExplorerPath} from '../pipelines/PipelinePathUtils';
 import {markdownToPlaintext} from '../ui/markdownToPlaintext';
 
+import {useAssetNodeMenu} from './AssetNodeMenu';
 import {buildAssetNodeStatusContent} from './AssetNodeStatusContent';
 import {AssetLatestRunSpinner} from './AssetRunLinking';
-import {LiveDataForNode} from './Utils';
+import {GraphData, GraphNode, LiveDataForNode} from './Utils';
 import {ASSET_NODE_NAME_MAX_LENGTH} from './layout';
 import {AssetNodeFragment} from './types/AssetNode.types';
 
@@ -36,7 +39,7 @@ export const AssetNode = React.memo(({definition, selected}: Props) => {
       <AssetTopTags definition={definition} liveData={liveData} />
       <AssetNodeContainer $selected={selected}>
         <AssetNodeBox $selected={selected} $isSource={isSource}>
-          <AssetName $isSource={isSource}>
+          <Name $isSource={isSource}>
             <span style={{marginTop: 1}}>
               <Icon name={isSource ? 'source_asset' : 'asset'} />
             </span>
@@ -50,14 +53,14 @@ export const AssetNode = React.memo(({definition, selected}: Props) => {
               })}
             </div>
             <div style={{flex: 1}} />
-          </AssetName>
+          </Name>
           <Box style={{padding: '6px 8px'}} flex={{direction: 'column', gap: 4}} border="top">
             {definition.description ? (
-              <AssetDescription $color={Colors.Gray800}>
+              <Description $color={Colors.Gray800}>
                 {markdownToPlaintext(definition.description).split('\n')[0]}
-              </AssetDescription>
+              </Description>
             ) : (
-              <AssetDescription $color={Colors.Gray400}>No description</AssetDescription>
+              <Description $color={Colors.Gray400}>No description</Description>
             )}
             {definition.isPartitioned && !definition.isSource && (
               <PartitionCountTags definition={definition} liveData={liveData} />
@@ -161,6 +164,91 @@ const AssetCheckIconsOrdered: {type: AssetCheckIconType; content: React.ReactNod
   },
 ];
 
+export const AssetNodeContextMenuWrapper = ({
+  children,
+  graphData,
+  explorerPath,
+  onChangeExplorerPath,
+  selectNode,
+  node,
+}: {
+  children: React.ReactNode;
+  graphData: GraphData;
+  node: GraphNode;
+  selectNode?: (e: React.MouseEvent<any> | React.KeyboardEvent<any>, nodeId: string) => void;
+  explorerPath?: ExplorerPath;
+  onChangeExplorerPath?: (path: ExplorerPath, mode: 'replace' | 'push') => void;
+}) => {
+  const [menuVisible, setMenuVisible] = React.useState(false);
+  const [menuPosition, setMenuPosition] = React.useState<{top: number; left: number}>({
+    top: 0,
+    left: 0,
+  });
+
+  const showMenu = (e: React.MouseEvent) => {
+    e.preventDefault();
+    setMenuVisible(true);
+    setMenuPosition({top: e.pageY, left: e.pageX});
+  };
+
+  const hideMenu = () => {
+    setMenuVisible(false);
+  };
+  const ref = React.useRef<HTMLDivElement | null>(null);
+  React.useEffect(() => {
+    const node = ref.current;
+    const listener = (e: MouseEvent) => {
+      if (ref.current && e.target && !ref.current.contains(e.target as Node)) {
+        hideMenu();
+      }
+    };
+    if (menuVisible && node) {
+      document.body.addEventListener('click', listener);
+    }
+    return () => {
+      if (node) {
+        document.body.removeEventListener('click', listener);
+      }
+    };
+  }, [menuVisible]);
+  const {dialog, menu} = useAssetNodeMenu({
+    graphData,
+    explorerPath,
+    onChangeExplorerPath,
+    selectNode,
+    node,
+  });
+  return (
+    <div ref={ref}>
+      <div onContextMenu={showMenu} onClick={hideMenu}>
+        {children}
+      </div>
+      {dialog}
+      {menuVisible
+        ? ReactDOM.createPortal(
+            <div
+              style={{
+                position: 'absolute',
+                top: menuPosition.top,
+                left: menuPosition.left,
+                backgroundColor: '#fff',
+                border: '1px solid #ccc',
+                boxShadow: '0 2px 4px rgba(0, 0, 0, 0.1)',
+                zIndex: 10,
+              }}
+              onClick={(e) => {
+                e.stopPropagation();
+              }}
+            >
+              {menu}
+            </div>,
+            document.body,
+          )
+        : null}
+    </div>
+  );
+};
+
 const AssetNodeChecksRow = ({
   definition,
   liveData,
@@ -239,9 +327,15 @@ export const AssetNodeMinimal = ({
             $background={background}
             $border={border}
           >
-            <AssetNodeSpinnerContainer>
+            <div
+              style={{
+                top: '50%',
+                position: 'absolute',
+                transform: 'translate(8px, -16px)',
+              }}
+            >
               <AssetLatestRunSpinner liveData={liveData} purpose="section" />
-            </AssetNodeSpinnerContainer>
+            </div>
             <MinimalName style={{fontSize: 30}} $isSource={isSource}>
               {withMiddleTruncation(displayName, {maxLength: 14})}
             </MinimalName>
@@ -328,7 +422,7 @@ const NameCSS: CSSObject = {
   fontWeight: 600,
 };
 
-export const NameTooltipCSS: CSSObject = {
+const NameTooltipCSS: CSSObject = {
   ...NameCSS,
   top: -9,
   left: -12,
@@ -347,19 +441,13 @@ const NameTooltipStyleSource = JSON.stringify({
   border: `1px solid ${Colors.Gray200}`,
 });
 
-const AssetName = styled.div<{$isSource: boolean}>`
+const Name = styled.div<{$isSource: boolean}>`
   ${NameCSS};
   display: flex;
   gap: 4px;
   background: ${(p) => (p.$isSource ? Colors.Gray100 : Colors.Blue50)};
   border-top-left-radius: 8px;
   border-top-right-radius: 8px;
-`;
-
-const AssetNodeSpinnerContainer = styled.div`
-  top: 50%;
-  position: absolute;
-  transform: translate(8px, -16px);
 `;
 
 const MinimalAssetNodeContainer = styled(AssetNodeContainer)`
@@ -395,7 +483,7 @@ const MinimalAssetNodeBox = styled.div<{
   }
 `;
 
-const MinimalName = styled(AssetName)`
+const MinimalName = styled(Name)`
   font-weight: 600;
   white-space: nowrap;
   position: absolute;
@@ -405,7 +493,7 @@ const MinimalName = styled(AssetName)`
   transform: translate(-50%, -50%);
 `;
 
-export const AssetDescription = styled.div<{$color: string}>`
+const Description = styled.div<{$color: string}>`
   white-space: nowrap;
   overflow: hidden;
   text-overflow: ellipsis;
